==> single_cell/primary_analysis/notebooks/01-setup.Rmd <==
### Paths and packages

If you are running this Rmd file within RStudio, make sure that "Knit Directory" is set to "Current Working Directory". The list of packages in the output **does not** include all of the attached or loaded packages within the session (e.g. `library(tidyverse)` loads in `dplyr`, `forcats`, `ggplot2`, `readr`, `purrr`, `stringr`, `tibble` and `tidyr`).

```{r setup}
# make sure the environment is clear of user objects
# note that custom set working directories, loaded libraries, etc. will be unaffected
# you may need to fully restart R for reproducibility
rm(list = ls())

# define the working directory (change as needed)
path_base <- file.path("~", "git", "Research", "LymeTCells", "single_cell")

# set the data directory (change as needed)
path_data_root <- file.path("/media", "edelaron", "Midas", "lyme")

# set other useful directories
path_objs <- file.path(path_data_root, "objects") # R (incl. Seurat) objects
path_resources <- file.path(path_base, "primary_analysis", "resources")
path_tables <- file.path(path_base, "primary_analysis", "tables")

# nf-core/airrflow specific
airrflow_version <- "4.1.0" # set the airrflow version that was run
path_airrflow_objs <-
  file.path(path_objs, "airr", paste0("airrflow_", airrflow_version))

# list (and install if needed) CRAN packages
packages <- c(# Immcantation packages
              "alakazam", "scoper", 
              # needed for manuscript figures
              "APackOfTheClones", "CellChat", "clusterProfiler", "ggpubr",
              "msigdbr", "SeuratExtend", "xlsx",
              # only for functions
              "circlize", "DESeq2", "DT", "ggsignif", "kableExtra", "knitr",
              "magrittr", "peakRAM", "rlang",
              # splicing
              "tximport",
              # plotting
              "ggrepel", "ggtext", "ggVennDiagram", "hues", "pals", "patchwork",
              # other
              "clustree", "here", "Matrix", "plyr", "reshape", "rjson",
              "scales", "Seurat", "tidyverse")
new_pkg <- packages[!(packages %in% installed.packages())]
if (length(new_pkg)) {install.packages(new_pkg)}

# list (and install if needed) Bioconductor packages
# have to have BiocManager already installed
packages_additional <- c("BiocParallel", "biomaRt", "celldex", "ComplexHeatmap",
                         "edgeR", "limma", "Nebulosa", "scDblFinder",
                         "scRNAseq", "SingleCellExperiment", "SingleR")
new_pkg_additional <-
  packages_additional[!(packages_additional %in% installed.packages())]
if (length(new_pkg_additional)) {BiocManager::install(new_pkg_additional)}

# load packages
packages <- sort(append(packages, packages_additional)) # so the list of versions will be in order
for (n in seq_along(packages)) {
  suppressPackageStartupMessages(library(packages[n], character.only = TRUE))
  # print simplified package versions
  cat(paste0(packages[n], ": ", packageVersion(packages[n]), "\n")) 
}

# remove unnecessary variables
rm(n, new_pkg, new_pkg_additional, packages, packages_additional)

# specify conflicting functions
filter <- dplyr::filter
select <- dplyr::select
```

### Cell types

Useful for pulling out immune cells later (e.g. for subclustering):

```{r annotations-cell-types}
# update these depending on your annotation style
t_cell_types <- "T Cells"

# for plot subtitles
t_cell_types_str <- str_c(t_cell_types, collapse = ", ")
```

### Visualization and themes

This will help the figures created later on to be more standardized and interpretable. Note that some colors have to be defined in later sections (several "Choose ... colors").

```{r colors}
# we had a dataset 1 that was not included in this manuscript
colors_dataset <- c("2" = "#47D1FF", "3" = "#0096C7", "4" = "#004D66") # shades of blue
colors_datatype <- c("Bulk" = "#788896", "GEX" = "#0CDD6B", "TCR" = "#007860") # shades of green

# for data type comparisons
colors_gex_tcr <- c("GEX" = colors_datatype[["GEX"]], "GEX_TCR" = "black",
                    "TCR" = colors_datatype[["TCR"]])

# tissue and sample types
colors_sample_type_blood <- c("Blood" = "#8B1A1A") # red
colors_sample_type_skin <- c("Control" = "#A6CEE3", "Control_EM" = "#00E321", 
                             "EM" = "#FF7F00") # pale blue, green, orange
colors_sample_types <- c(colors_sample_type_blood, colors_sample_type_skin)

# QC specific
colors_chain_tcr <- c("TRBOnly" = "#007F5F", "Paired" = "#55A630",
                      "TRAOnly" = "#DDDF00") # yellow and green

# TRUE vs. FALSE
colors_binary <- c("TRUE" = "#3FEBC5", "FALSE" = "#C2B497") # turquoise and sand

# for doublet plots
colors_doublet <- c("Singlet" = "#cbcae3", "Doublet" = "#807DBA")

# for clustrees (to make the clusters easier to distinguish)
colors_clustree <- setNames(stepped2(n = 16),
                            nm = seq(0.1, 1.6, 0.1)) # might need adjusting

# manuscript
colors_immune <- c("Adaptive Immune" = "#006358", "Innate Immune" = "#00B19D",
                   "Immune" = "#008A7B", "Non Immune" = "#3ED55D")
```

These functions aid with ggplot2 plotting:

```{r ggplot-themes}
# yes I know that I could have picked a name that wouldn't seem misspelled
theme_bw <- theme_bw() +
              theme(panel.grid.major.x = element_blank(),
                    panel.grid.major.y =
                      element_line(color = "grey75", linewidth = 0.2))

labels_rotate_x <- theme(axis.text.x = element_text(angle = 45, hjust = 1))

# having strip.text.x = element_text(size = 7) messes with add_info_bar
labels_standard <- theme(plot.title = element_text(size = 12, hjust = 0.5),
                         plot.subtitle = element_text(size = 9, hjust = 0.5),
                         plot.caption = element_text(size = 7),
                         axis.title = element_text(size = 9),
                         axis.text = element_text(size = 8),
                         legend.title = element_text(size = 10),
                         legend.text = element_text(size = 8))

labels_standard_venn <- theme(plot.title = element_text(size = 12, hjust = 0.5),
                              plot.subtitle =
                                element_text(size = 9, hjust = 0.5),
                              legend.title = element_text(size = 10),
                              legend.text = element_text(size = 10))

# removes "Identity"
labels_standard_vln <- labels_standard +
                       theme(plot.title = element_text(size = 14, hjust = 0.5,
                                                       face = "plain"),
                             axis.title.x = element_blank(),
                             axis.title.y = element_text(size = 10),
                             axis.text = element_text(size = 10))

# the label is a little misleading; this rotates the labels back to flat
labels_standard_vln_rotate <- labels_standard_vln +
                              theme(axis.text.x =
                                      element_text(angle = 0, hjust = 0.5))

# just for clustrees
labels_clustree <- theme(axis.title = element_blank(),
                         axis.text = element_blank(),
                         axis.ticks = element_blank())

# sometimes just axis.text doesn't do anything
# having an aspect ratio messes up patchwork sometimes
clean_umap <- theme(plot.title = element_text(hjust = 0.5, face = "plain"),
                    axis.text = element_blank(), axis.text.x = element_blank(),
                    axis.text.y = element_blank(), axis.ticks = element_blank(),
                    aspect.ratio = 1)

legend_bottom <- theme(legend.position = "bottom",
                       legend.justification = "center",
                       legend.title = element_blank()) # no title - they mess up spacing

# for patchwork plots
plot_anno <- plot_annotation(tag_levels = "A")
```

More plot settings:

```{r plot-options}
# so that the UMAPs and other plots stay consistent
# note that this will affect anything with pseudorandomness
set.seed(42)

# to suppress a warning message
options(ggrepel.max.overlaps = Inf)
```

### Features to be filtered out

You shouldn't just pull out genes that begin with "IG" or "TR"

The following show an example of the Ensembl 93 features, the gene biotypes that they use and an example of the genes that we exclude:

```{r filter-features-load}
#| warning = FALSE

# load this when needed e.g. to run seurat_pipeline()

# read in meta data pulled from Ensembl 93 using biomaRt
load(file.path(path_resources, "QC_features_meta.RData")) # features_meta, features_ensembl, features_readable

# there are some genes that start with "IG" or "TR" that are not immune-related,
# so you shouldn't just remove genes based upon that
biotypes_excl <-
  unique(features_meta[["gene_biotype"]])[grepl(pattern = "^IG_|^TR_",
                                                x = unique(features_meta[["gene_biotype"]]))]
remove_genes <- features_meta %>%
                  dplyr::filter(gene_biotype %in% biotypes_excl) %>%
                  pull(external_gene_name)

# remove unnecessary variables
# rm(biotypes_excl, ensembl, features_meta, mart)
```

If you want to download an Ensembl database yourself, you can use this code:

```{r}
#| eval = FALSE

# credit to Edward Lee and Hailong Meng in the Kleinstein lab
# make sure that you have biomaRt and tidyverse loaded

# define your species and genome of interest (use listDatasets below for other choices)
species <- "human" # or "mouse"
genome <- "hsapiens" # or "mmusculus"

# get the current version
current_version <- listEnsemblArchives() %>%
                     dplyr::filter(current_release == "*") %>%
                     pull(name)
cat(paste("The current version is:", current_version))
current_version <- str_split(current_version, " ")[[1]][2] # just the number
# or pick an older version

# search Ensembl's datasets
ensembl <- useEnsembl(biomart = "ensembl")
dataset <- searchDatasets(mart = ensembl, pattern = genome)$dataset

# use listEnsembl() for biomart options (we always use genes)
# you could set mirror = "" but you'd have to remove the version parameter
ensembl <- useEnsembl(biomart = "genes", dataset = dataset,
                      version = current_version)
# could also just do:
# ensembl <- useMart(biomart = "ensembl", dataset = dataset)

# filter for the features of interest
filters <- listFilters(ensembl)
attributes <- listAttributes(ensembl)
attributes <- dplyr::filter(attributes,
                            name %in% c("ensembl_gene_id", "external_gene_name",
                                        "gene_biotype", "hgnc_symbol",
                                        "description") & page == "feature_page")

# known biotypes
tr_ig_biotypes <- c("TR_C_gene", "TR_D_gene",
                    "TR_J_gene", "TR_J_pseudogene",
                    "TR_V_gene", "TR_V_pseudogene",
                    "IG_C_gene", "IG_C_pseudogene",
                    "IG_D_gene", "IG_D_pseudogene",
                    "IG_J_gene", "IG_LV_gene", "IG_pseudogene",
                    "IG_V_gene", "IG_V_pseudogene")

features_meta <- getBM(attributes = c("ensembl_gene_id", "external_gene_name",
                                      "gene_biotype", "hgnc_symbol",
                                      "description"),
                       filters = "biotype", values = tr_ig_biotypes,
                       mart = ensembl)

# save the file (change as desired)
file_name <- paste0("QC_features_meta_", species, "_v", current_version, ".rds")
write_rds(features_meta, file = file.path(path_resources, file_name))
```

Examples of genes to be excluded:

```{r filter-features-example}
#| eval = FALSE

kable(head(features_meta, 10)) %>%
  kable_styling("striped") %>%
  scroll_box(height = NULL, width = "100%")

# feature biotypes (want to exclude IGs and TRs)
kable(table(features_meta[["gene_biotype"]])) %>%
  kable_styling("striped") %>%
  scroll_box(height = NULL, width = "100%")

head(remove_genes, 20)
```
